tests/wpt/web-platform-tests/web-bundle/subresource-loading/subresource-loading-nested-bundle.tentative.html

<!DOCTYPE html>
<title>A nested bundle is not supported</title>
<link
  rel="help"
  href="https://github.com/WICG/webpackage/blob/master/explainers/subresource-loading.md"
/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <link
    rel="webbundle"
    href="http://web-platform.test:8001/web-bundle/resources/wbn/nested-main.wbn"
    resources="http://web-platform.test:8001/web-bundle/resources/wbn/nested-sub.wbn"
  />
  <script>
    promise_test(async () => {
      const response = await fetch(
        "http://web-platform.test:8001/web-bundle/resources/wbn/nested-sub.wbn"
      );
      assert_true(response.ok);
    }, "A nested bundle can be fetched");

    promise_test(async () => {
      await addLinkAndWaitForError(
        "http://web-platform.test:8001/web-bundle/resources/wbn/nested-sub.wbn",
        "http://web-platform.test:8001/web-bundle/resources/wbn/root.js"
      );
      const response = await fetch(
        "http://web-platform.test:8001/web-bundle/resources/wbn/root.js"
      );
      assert_false(response.ok);
    }, "Subresources in a nested bundle should not be loaded");

    function addLinkAndWaitForError(url, resources) {
      return new Promise((resolve, reject) => {
        const link = document.createElement("link");
        link.rel = "webbundle";
        link.href = url;
        link.resources = resources;
        link.onload = reject;
        link.onerror = () => resolve();
        document.body.appendChild(link);
      });
    }
  </script>
</body>
